Create `PlantEvents` during import
Deleting the `PlantInput` type from the GraphQL schema will cause Fauna to create the default `PlantInput` type which also exposes the `events` relation. We then use this to create events at the same time as creating a `Plant`.

Unfortunately this means that the `updatePlant()` function will now allow the caller to create new events.

Rather than fight any more with Fauna to get this to work how I want I'm going to just let this be and in the next task move us to AWS AppSync.