Create `GraphqlResult`
Amplify has its own type representing a GraphQL result but it combines the success and error types making it impossible to distinguish between them without putting null checks all over the place.

These types are compatible (i.e. have all the same properties) but implement the constraint that a response can be either successful or erroneous but not both.