Add generated TypeScript
So that CI doesn't need to generate the code. It would be better if it could, but it
requires checking out an Amplify environment (before `amplify codegen` will run),
which means CI would need to know _which_ environment to use, which raises a whole
host of questions.

This way CI can build independently of Amplify environments.