Clone plant events array instead of reusing reference
When we update a plant the server (really the generated GraphQL query) returns only the plant attributes, not its nested events collection. We add the `events` array into the returned object to match the `Plant` type definition and previously we simply copied over the reference to the original `events` array (from the `Plant` object before the update).

Because the events array is reused, changes to it will affect any plant with a reference to the array, which may have unintended consequences.

In this commit we change the update logic to make a copy of the `events` array using lodash's `cloneDeep()` (thanks @reydelo for the suggestion)