Tie the plant conversion functions to GraphQL structures
The original `plantToGraphqlPlant()` function became out of date when new, required fields were added to the GraphQL `Plant` structure.

Here we tie the return value of the conversion functions to the GraphQL generated types so that in future if the generated types change, the conversion functions will break, and the changes to the function inputs will propagate out to the places they're used as well.